+++ basxconnect/core/layouts/editperson_legal.py
@@ -2,17 +2,18 @@ import htmlgenerator as hg
 from bread import layout
 from django.utils.translation import gettext_lazy as _
 
-from basxconnect.core.layouts.person import (
-    R,
-    C,
-    F,
+from basxconnect.core.layouts.editperson import (
     contact_details,
     relationshipstab,
     style_person,
 )
 
+R = layout.grid.Row
+C = layout.grid.Col
+F = layout.form.FormField
 
-def personassociation(request):
+
+def editperson_legal_tabs(request):
     # fix: alignment of tab content and tab should be on global grid I think
     ret = layout.tabs.Tabs(
         base_data_tab(request),
@@ -28,22 +29,27 @@ def base_data_tab(request):
         _("Base data"),
         hg.BaseElement(
             layout.grid.Grid(
-                R(C(hg.H4(_("General Information")))),
                 R(
                     C(
+                        R(C(hg.H4(_("Name")))),
                         R(
-                            C(F("name")),
-                            C(),
+                            C(
+                                R(C(F("name"))),
+                                R(C(F("name_addition"))),
+                                width=8,
+                                breakpoint="lg",
+                            )
                         ),
-                        R(
-                            C(),
-                            C(F("preferred_language")),
-                        ),
+                        _class="section-separator-right",
                     ),
                     C(
+                        R(C(hg.H4(_("Mailings")))),
                         R(
-                            C(),
-                            C(F("salutation_letter")),
+                            C(F("preferred_language"), width=4, breakpoint="lg"),
+                            C(F("type"), width=8, breakpoint="lg"),
+                        ),
+                        R(
+                            C(F("salutation_letter"), width=12, breakpoint="lg"),
                         ),
                     ),
                 ),
